Updated FastAPI

+++ task_management_app/src/tasks/controller.py
@@ -2,12 +2,15 @@ from src.tasks.dtos import TaskSchema
 from sqlalchemy.orm import Session
 from src.tasks.models import TaskModel
 from fastapi import HTTPException
+from src.user.models import UserModel
 
-def create_task(body: TaskSchema, db:Session):
+
+def create_task(body: TaskSchema, db: Session, user: UserModel):
     data = body.model_dump()
-    new_task = TaskModel(title = data["title"],
-                         description = data["description"],
-                         is_completed = data["is_completed"]
+    new_task = TaskModel(title=data["title"],
+                         description=data["description"],
+                         is_completed=data["is_completed"],
+                         user_id=user.id
                          )
     db.add(new_task)
     db.commit()
@@ -15,13 +18,12 @@ def create_task(body: TaskSchema, db:Session):
     return new_task
 
 
-def get_tasks(db:Session):
-    tasks = db.query(TaskModel).all()
+def get_tasks(db: Session, user: UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 
-
-def get_one_task( task_id:int, db:Session):
+def get_one_task(task_id: int, db: Session):
     one_task = db.query(TaskModel).filter(TaskModel.id == task_id).first()
 
     if not one_task:
@@ -29,10 +31,14 @@ def get_one_task( task_id:int, db:Session):
 
     return one_task
 
-def update_task(body:TaskSchema, task_id:int, db:Session):
-    one_task = db.query(TaskModel).filter(TaskModel.id == task_id).first()
+
+def update_task(body: TaskSchema, task_id: int, db: Session, user: UserModel):
+    one_task: TaskModel = db.query(TaskModel).filter(TaskModel.id == task_id).first()
     if not one_task:
         raise HTTPException(status_code=404, detail="Task Id is Incorrect")
+
+    if one_task.user_id != user.id:
+        raise HTTPException(status_code=401, detail="You are not allowed to update this task")
 
     body = body.model_dump()
     for field, value in body.items():
@@ -45,15 +51,15 @@ def update_task(body:TaskSchema, task_id:int, db:Session):
     return one_task
 
 
-
-def delete_task(task_id:int, db:Session):
+def delete_task(task_id: int, db: Session, user: UserModel):
     one_task = db.query(TaskModel).filter(TaskModel.id == task_id).first()
     if not one_task:
         raise HTTPException(status_code=404, detail="Task Id is Incorrect")
+
+    if one_task.user_id != user.id:
+        raise HTTPException(status_code=401, detail="You are not allowed to delete this task")
 
     db.delete(one_task)
     db.commit()
 
     return None
-
-
